+++ packages/component-library/src/components/templates/ApplicationLayout/ApplicationLayout.css.ts
@@ -96,7 +96,7 @@ globalStyle('header a,footer a', {
   cursor: 'pointer',
 });
 
-globalStyle('h1', globalHeadingStyle({ level: '1', weight: 'medium' }));
+globalStyle('h1', { ...globalHeadingStyle({ level: '1', weight: 'medium' }) });
 globalStyle('h2', globalHeadingStyle({ level: '2', weight: 'medium' }));
 globalStyle('h3', globalHeadingStyle({ level: '3', weight: 'medium' }));
 globalStyle('h4', globalHeadingStyle({ level: '4', weight: 'medium' }));
@@ -108,10 +108,6 @@ globalStyle('h1', {
 globalStyle('h2,h3,h4', {
   marginTop: vars.space['2x'],
   marginBottom: vars.space['2x'],
-});
-
-globalStyle('h1,h2,h3,h4', {
-  color: vars.headings.color,
 });
 
 globalStyle('p', {