+++ packages/website/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cutting/website
 
+## 4.27.9
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @cutting/component-library@5.3.4
+
 ## 4.27.8
 
 ### Patch Changes